re-organize core module
